Added magnetic torque mechanism
Added magnetic torque mechanism to rotate the player towards the
direction of the surface attracted by the magnet. See the Magnetic
Torque scene for demo.

Removed programmatic implementation of player rigidbody in
OpenPhysicsObject, as we'll add the rigidbody though the Unity designer
going forward. I implemented this change in the Magnetic Torque scene
and Applied it to the Dan Character prefab.